Fix: Required item update
I've been having this issue for a good month now. Looked through the script, noticed this event doesn't match the event in QBCore. Updating this to match actually updates the PlayerData so it recognizes when an item is in your inventory or broken, dropped, etc. Tested in a solo dev environment; is also the same method used in nui_doorlock.

Fixes #56 and maybe others