refactor: qtarget changes
- Fixed updating zones not working because updating the whole zone table breaks it
- Fix wrong coords on bones checker in loop
- Fix oal not being enabled due to the new name for the flag being implemented
- Replaced the raycast flag and added a los check:
"Some entities aren't hit by raycasts when interacting with the world,
which is why we alternate flags. When using flag 30, check line-of-sight
to ensure it's not being hit through a wall.
Reduced raycast length from 10000 to 16, and collider to 7.

Didn't notice anything break from this, but needs confirmation."
- Linden

- EnableTarget runs in a thread now, it needs to because a command function isn't in a thread already so calling yielding functions (for example wait) will cause silent errors
- Support objects with invalid types:
"Some entities will return an entity type of 0 ("no entity"), despite clearly
being an entity. Calling GetEntityModel either returns a model, or throws
a native invocation error.
Use a protected call for GetEntityModel and treat anything with a model
as an object.

Notably fixes interaction with (most?) trees, who knows what else."
- Linden