fix(client): Fix inconsistently targeting models
Switches the flag for `StartShapeTestLosProbe` from `7` to `4` to fix weird issues with inconsistent target models. 

Upgrading from 5.3.9=>5.4.0, many models were unable to be targeted, mostly identified immediately with road signs. After debugging thoroughly, I identified this change. Changing the bitmask to 4 fixes the issues encountered and all expected targets were able to be targeted. I also tried accessing a target inside of another target (storage container), and sitting in a chair through a wall - was unable to do either as expected.

This may resolve #158 if it's still an active issue. I tried to replicate but was unable to.

After further review, I noticed changing the bitmask to 4 was also in the upstream qtarget.